Switch from float to CGFloat for column origins in table view
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/gui/branches/gnustep_testplant_branch@40255 72102866-910b-0410-8b05-ffd578937521

+++ Headers/AppKit/NSTableView.h
@@ -157,7 +157,7 @@ typedef enum _NSTableViewAnimationOptions
    * origin of each column).  When a column width is changed through
    * [NSTableColumn setWidth:], then [NSTableView tile] gets called,
    * which updates the cache.  */
-  float *_columnOrigins;
+  CGFloat *_columnOrigins;
 
   /*
    *  We keep the superview's width in order to know when to
